backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
openai>=1.12.0
yt-dlp>=2026.1.29
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.25.0,<0.28
celery>=5.3.6
redis>=5.0.1
supabase>=2.3.4
google-genai>=1.0.0
playwright>=1.40.0
playwright-stealth>=2.0.0
